setting for game
